Remove abuse use of int8_t types in cc1
It is desirable this compiler can be executed in real z80 machines
and it is the reason why there were a lot of variables with int8_t
types, but it caused the code to be non idiomatic at all. This
patch changes all these types to int/unsigned, with the convention
they must hold values of int8_t/uint8_t types, so an option of the
compiler can generate code with 8 bit integers and have the same
result. This same option is present in the AVR gcc port.

+++ cc1/error.c
@@ -9,11 +9,11 @@
 
 #define MAXERRNUM 10
 
-extern uint8_t failure;
-static uint8_t nerrors;
+extern int failure;
+static unsigned nerrors;
 
 static void
-warn_helper(int8_t flag, char *fmt, va_list va)
+warn_helper(int flag, char *fmt, va_list va)
 {
 	if (!flag)
 		return;
@@ -30,7 +30,7 @@ warn_helper(int8_t flag, char *fmt, va_list va)
 void
 warn(char *fmt, ...)
 {
-	extern uint8_t warnings;
+	extern int warnings;
 
 	va_list va;
 	va_start(va, fmt);
